Refactor Android cloudsync instrumentation tests
